Make `Promise.cast` work on subclasses.
It first checks that `x.[[IsPromise]]`, which will be true for all `Promise`s and for all ES6 `Promise`-subclass instances. Then, it ensures that it only passes through promises for which `Object.getPrototypeOf(promise).constructor === this`, so that `Promise.cast` only passes through promises whose proto's constructor is `Promise`, whereas `CancellablePromise.cast` only passes through promises whose proto's constructor is `CancellablePromise.

To secure this, freeze the `Promise.prototype.constructor` property. Note that because of the `[[GetInheritance]]`, shadowing a derived promise's `constructor` property will not suffice to fool the second check.