Make `then` work with subclasses.
It now automatically returns an instance of the subclass. Effectively, it constructs it via `new this.constructor(() => {})` and then uses internal magic to resolve or reject the returned promise. This is unlike the static methods which save `resolve` and `reject` from the resolver to perform their work. But this seems OK, as `then` is so fundamental, that you should not be able to add new behavior to it by passing in modified `resolve`/`reject` arguments to the resolver in your subclass.